requirements.txt
Django==1.9.5
djangorestframework
markdown
django-filter